refactor: Rename package and update CostCenter related methods and mappings
Rename `custcenter` package to `cost_center`. Update mappedBy from `cost_center` to `costCenter` in CostCenter model. Rename `getCostCenterResponseDto` method to `getResponseDto` in CostCenterService. Update `RequestMapping` from `/api/cost-center` to `/api/cost-centers` in CostCenterController.
