+++ src/main/java/com/example/spending/dto/cost_center/CostCenterResponseDto.java
@@ -1,3 +1,3 @@
-package com.example.spending.dto.custcenter;
+package com.example.spending.dto.cost_center;
 
 public class CostCenterResponseDto extends CostCenterRequestDto {}
